requirements.txt
Django==3.0.5
psycopg2==2.7.7
django-cors-headers==3.2.1 
djangorestframework==3.11.0